Fill out Usage section
While I could list all of the available methods, I think that the GoDocs already do this fairly well and I would be duplicating data. However, a tracking issue with all of the methods that are needed would be ace. Shall I set one of those up?

I also point to the spec here. If someone points me to a Go package with a good Usage and API section, I am happy to fill this out more; however, I am not sure how else to best go forth with filling this section out.