Added tests, cleaned up publish code, added additional function
At the expense of slight duplicated code, I create two new functions getURLIPNS and SendIPNS to accommodate for how the daemon handles web requests. Currently when using `Send` every argument gets formatted with `arg=....&arg=....` causing defaults to be used for many arguments. This was preventing publishing of records using non-self keys as the arguments for the API calls have to be named.