travis: enable stream mounting and configure IPFS listeners
We need to enable the IPFS feature Experimental.Libp2pStreamMounting
in order to test the P2P APIs. To do so, we have to run `ipfs init`
manually so that the IPFS config file exists on disk. In addition, we
also have to configure Addresses.API and Addresses.Gateway, which were
previously set by the container at boot, because the container stops
setting those parameters if the IPFS configuration file already
exists.