remove the go-libp2p-pubsub dependency
We don't actually need to use the struct definitions from go-libp2p-pubsub. The
interface is part of the API so we should be able to just re-implement them here.

This *is* a breaking change: I've removed the interface and replaced it with a
bare struct. IMO, that's *significantly* more idiomatic. `SeqNo()` needed to be
changed *anyways* as we can't guarantee that sequence numbers are *just*
numbers, they can be anything. The existing code would have crashed when run
against js-ipfs, as far as I could tell.